Move ConcertEvent into example.application package
