+++ sample-app/src/test/scala/example/application/projection/DatabaseConcertProjectionRepositoryBehaviors.scala
@@ -2,7 +2,7 @@ package example.application.projection
 
 import example.DatabaseSpecBase
 import example.adapter.{ ConcertId, ConcertIdGenerator, ConcertTicketId }
-import example.adapter.ConcertEvent.{ ConcertCancelled, ConcertCreated, ConcertTicketsBought }
+import example.application.ConcertEvent.{ ConcertCancelled, ConcertCreated, ConcertTicketsBought }
 import example.readmodel.ConcertDatabaseService
 import slick.dbio.DBIO
